종양심기. 아직 충분히 테스트 안해봄.

+++ Assets/Script/Object/Plant.cs
@@ -14,10 +14,11 @@ public class Plant : MonoBehaviour
     public bool isGatherPossible;
     public string plantName = "StickPlant";
     public int GatherAnimationType = 1;
-    public int state; //0:자라는 중(덫설치후 대기중), 1:채집가능(덫잡힘), 2:채집후, 3:덫설치중
+    public int state; //0:자라는 중(덫설치후 대기중), 1:채집가능(덫잡힘), 2:채집후, 3:덫설치중, 4:종양심음, 5:종양채집가능
 
     public float growthTime; //성장하는데 걸리는 시간(초)
     public float growthTimer = 0;
+    public bool isTumor = false;
 
     void Start ()
     {
@@ -40,29 +41,48 @@ public class Plant : MonoBehaviour
 
     public void GetItem()
     {
-        switch (plantName)
+        if (isTumor == false)
         {
-            case "StickPlant":
-                inventory.GetItem(global::Inventory.Item.Stick, 2);
-                break;
-            case "MassPlant":
-                inventory.GetItem(global::Inventory.Item.Mass, 1);
-                SceneObjectManager.instance.DeleteObject(sceneNum, Grid.instance.PosToGrid(transform.position.x));
-                break;
-            case "BoardPlant":
-                inventory.GetItem(global::Inventory.Item.Board, 1);
-                break;
-            case "ThornPlant":
-                inventory.GetItem(global::Inventory.Item.Thorn, 5);
-                break;
-            case "Trap01":
-                inventory.GetItem(global::Inventory.Item.Hose, 1);
-                inventory.GetItem(global::Inventory.Item.Heart, 1);
-                inventory.GetItem(global::Inventory.Item.Mass, 1);
-                SceneObjectManager.instance.DeleteObject(sceneNum, Grid.instance.PosToGrid(transform.position.x));
-                break;
+            switch (plantName)
+            {
+                case "StickPlant":
+                    inventory.GetItem(global::Inventory.Item.Stick, 2);
+                    break;
+                case "MassPlant":
+                    inventory.GetItem(global::Inventory.Item.Mass, 1);
+                    SceneObjectManager.instance.DeleteObject(sceneNum, Grid.instance.PosToGrid(transform.position.x));
+                    break;
+                case "BoardPlant":
+                    inventory.GetItem(global::Inventory.Item.Board, 1);
+                    break;
+                case "ThornPlant":
+                    inventory.GetItem(global::Inventory.Item.Thorn, 5);
+                    break;
+                case "Trap01":
+                    inventory.GetItem(global::Inventory.Item.Hose, 1);
+                    inventory.GetItem(global::Inventory.Item.Heart, 1);
+                    inventory.GetItem(global::Inventory.Item.Mass, 1);
+                    SceneObjectManager.instance.DeleteObject(sceneNum, Grid.instance.PosToGrid(transform.position.x));
+                    break;
+            }
+        }
+        else
+        {
+            switch (plantName)
+            {
+                case "StickPlant":
+                    inventory.GetItem(global::Inventory.Item.Tumor, 2);
+                    break;
+                case "BoardPlant":
+                    inventory.GetItem(global::Inventory.Item.Tumor, 2);
+                    break;
+                case "ThornPlant":
+                    inventory.GetItem(global::Inventory.Item.Tumor, 6);
+                    break;
+            }
         }
         isGatherPossible = false;
+        isTumor = false;
 
         if (Grid.instance.PosToGrid(transform.position.x) == Grid.instance.PlayerGrid())
         {
@@ -79,25 +99,54 @@ public class Plant : MonoBehaviour
     public bool InventoryCheck()
     {
         bool temp = true;
-        switch (plantName)
+        if (isTumor == false)
         {
-            case "StickPlant":
-                temp = !inventory.isFull(1, global::Inventory.Item.Stick, 2);
-                break;
-            case "MassPlant":
-                temp = !inventory.isFull(1, global::Inventory.Item.Mass, 1);
-                break;
-            case "BoardPlant":
-                temp = !inventory.isFull(1, global::Inventory.Item.Board, 1);
-                break;
-            case "ThornPlant":
-                temp = !inventory.isFull(1, global::Inventory.Item.Thorn, 5);
-                break;
-            case "Trap01":
-                temp = !inventory.isFull(2, global::Inventory.Item.Hose, 1, global::Inventory.Item.Heart, 1, global::Inventory.Item.Mass, 1);
-                break;
+            switch (plantName)
+            {
+                case "StickPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Stick, 2);
+                    break;
+                case "MassPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Mass, 1);
+                    break;
+                case "BoardPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Board, 1);
+                    break;
+                case "ThornPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Thorn, 5);
+                    break;
+                case "Trap01":
+                    temp = !inventory.isFull(3, global::Inventory.Item.Hose, 1, global::Inventory.Item.Heart, 1, global::Inventory.Item.Mass, 1);
+                    break;
+            }
+        }
+        else
+        {
+            switch (plantName)
+            {
+                case "StickPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Tumor, 2);
+                    break;
+                case "BoardPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Tumor, 2);
+                    break;
+                case "ThornPlant":
+                    temp = !inventory.isFull(1, global::Inventory.Item.Tumor, 6);
+                    break;
+            }
         }
         return temp;
+    }
+
+    public void SetTumor()
+    {
+        if (state == 2 && (plantName == "StickPlant" || plantName == "BoardPlant" || plantName == "ThornPlant"))
+        {
+            state = 4;
+            animaitor.SetInteger("State", state);
+            interactionIcon.DeleteIcon(global::InteractionIcon.Icon.Tumor);
+            inventory.DeleteItem(Inventory.Item.TumorSeed);
+        }
     }
 
     //애니메이션 이벤트에서 사용하는 함수
@@ -106,6 +155,13 @@ public class Plant : MonoBehaviour
         isGatherPossible = false;
         if (Grid.instance.PosToGrid(transform.position.x) == Grid.instance.PlayerGrid())
         {
+            if (state == 2 && (plantName == "StickPlant" || plantName == "BoardPlant" || plantName == "ThornPlant"))
+            {
+                if (inventory.HasItem(Inventory.Item.TumorSeed) == true)
+                {
+                    interactionIcon.AddIcon(global::InteractionIcon.Icon.Tumor);
+                }
+            }
             interactionIcon.AddIcon(global::InteractionIcon.Icon.Remove);
         }
     }
@@ -133,11 +189,16 @@ public class Plant : MonoBehaviour
         {
             Growth();
         }
+        if(state == 4 || state == 5)
+        {
+            isTumor = true;
+        }
+        animaitor.SetInteger("State", state);
     }
 
     void Growth()
     {
-        if (state == 0 || state == 2)
+        if (state == 0 || state == 2 || state == 4)
         {
             if (plantName == "Trap01")
             {
@@ -163,12 +224,19 @@ public class Plant : MonoBehaviour
             if (growthTimer >= growthTime)
             {
                 growthTimer = 0;
-                state = 1;
+                if(state == 4)
+                {
+                    state = 5;
+                }
+                else
+                {
+                    state = 1;
+                }
                 animaitor.SetInteger("State", state);
             }
         }
 
-        if (state == 1 && growthTimer != 0)
+        if ((state == 1 || state == 5) && growthTimer != 0)
         {
             growthTimer = 0;
         }
@@ -178,10 +246,18 @@ public class Plant : MonoBehaviour
     {
         if (other.gameObject.tag == "Player" && inventory.isInventoryActive == false)
         {
+            if(state == 2 && (plantName == "StickPlant" || plantName == "BoardPlant" || plantName == "ThornPlant"))
+            {
+                if (inventory.HasItem(Inventory.Item.TumorSeed) == true)
+                {
+                    interactionIcon.AddIcon(global::InteractionIcon.Icon.Tumor);
+                }
+            }
             if (isGatherPossible == true)
             {
                 interactionIcon.AddIcon(global::InteractionIcon.Icon.Gather);
                 interactionIcon.DeleteIcon(global::InteractionIcon.Icon.Remove);
+                interactionIcon.DeleteIcon(global::InteractionIcon.Icon.Tumor);
             }
             else
             {
@@ -197,6 +273,7 @@ public class Plant : MonoBehaviour
         {
             interactionIcon.DeleteIcon(global::InteractionIcon.Icon.Gather);
             interactionIcon.DeleteIcon(global::InteractionIcon.Icon.Remove);
+            interactionIcon.DeleteIcon(global::InteractionIcon.Icon.Tumor);
         }
     }
 }
